refactor: team Schedule domain
