+++ src/repositories/teamScheduleRepository.js
@@ -53,4 +53,32 @@ export const teamScheduleRepository = {
       data: scheduleData,
     });
   },
+
+  getTeamMemberUserIds: async (teamId) => {
+    return await prisma.teamMember.findMany({
+      where: { team_id: Number(teamId) },
+      select: { user_id: true },
+    });
+  },
+  getUserSchedules: async (userIds, startDate, endDate) => {
+    return await prisma.userSchedule.findMany({
+      where: {
+        user_id: { in: userIds },
+        user_schedule_start: { gte: startDate, lte: endDate },
+      },
+      select: {
+        user_id: true,
+        user_schedule_start: true,
+        user_schedule_end: true,
+      },
+      orderBy: { user_schedule_start: "asc" },
+    });
+  },
+
+  getUserNames: async (userIds) => {
+    return await prisma.user.findMany({
+      where: { user_id: { in: userIds } },
+      select: { user_id: true, user_name: true },
+    });
+  },
 };
